+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [8.11.0]
+
+### Added
+
+- E2E tests for the forms.
+- Date field now supports `multiple` and `range` modes.
+
+### Removed
+
+- Old outdated tests.
+
 ## [8.10.2]
 
 ### Fixed
@@ -1614,6 +1625,8 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 - Initial production release.
 
+
+[8.11.0]: https://github.com/infinum/eightshift-forms/compare/8.10.2...8.11.0
 [8.10.2]: https://github.com/infinum/eightshift-forms/compare/8.10.1...8.10.2
 [8.10.1]: https://github.com/infinum/eightshift-forms/compare/8.10.0...8.10.1
 [8.10.0]: https://github.com/infinum/eightshift-forms/compare/8.9.3...8.10.0